fix: accept empty ref/alt in report variant parsing
extract local SPARQList proxy logic into
app/frontend/config/sparqlistProxy.js
and import it in webpack configs; pass endpoint to dev proxy creator
